[#IOPID-1926] Refactor types and config
